% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwig.R
\name{get_names}
\alias{get_names}
\title{Get HWI/HWIG names}
\usage{
get_names(DT, by)
}
\arguments{
\item{DT}{input group membership data, in individual/group format}

\item{by}{column(s) to split calculation by. e.g.: year}
}
\value{
names corresponding to values of by for each of the returned list of matrices in `calc_hwi` and `calc_hwig`.
}
\description{
Helper function, to return names of each matrix
}
\examples{
data(DT)
hwi <- calc_hwi(DT, 'id', 'group', 'yr')
hwig <- calc_hwig(hwi)
nms <- get_names(DT, 'yr')
names(hwig) <- nms
}
