% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwig.R
\name{calc_hwi}
\alias{calc_hwi}
\title{Calculate HWI}
\usage{
calc_hwi(ind.groups, input.data.format = "individuals",
  network.data.format = "GBI", assc.ind = "HWI")
}
\arguments{
\item{ind.groups}{input group membership data, in ID/GROUP format}

\item{input.data.format}{groups or individuals, as string (for get_group_by..)}

\item{network.data.format}{GBI or SP, group by individual or sampling period}

\item{assc.ind}{SRI or HWI, for association_index}
}
\value{
data.table of HWI for each individual pair combination
}
\description{
Calculate HWI
}
