% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwig.R
\name{calc_hwi}
\alias{calc_hwi}
\title{Calculate HWI}
\usage{
calc_hwi(DT, id, group)
}
\arguments{
\item{DT}{input group membership data, in individual/group format}

\item{id}{column indicating id in DT}

\item{group}{column indicating group in DT}
}
\value{
data.table of HWI
}
\description{
Calculate HWI
}
